crewai-cursor-agent/requirements.txt
crewai==0.16.2
anthropic==0.38.0
streamlit==1.39.0
langchain==0.1.12
langchain-anthropic==0.1.4
langchain-core==0.1.32
pandas==2.2.3
numpy>=1.26.4
python-pptx==0.6.23
PyPDF2==3.0.1
pillow==10.4.0
firecrawl==1.4.0
duckduckgo-search==4.5.0
httpx>=0.23.3,<0.25.0
matplotlib==3.8.3
python-dotenv==1.0.1
pytest==8.0.2
pylint==3.0.3
flake8==7.0.0
google-auth==2.23.4
google-auth-httplib2==0.1.0
httplib2>=0.22.0
grpcio==1.54.2
protobuf>=4.21.6
proto-plus>=1.22.3
rich>=13.7.0,<14.0.0
msgpack>=1.0.5 